separated security registration into new SecurityPlugin and instead made WebArchitecturePlugin a meta-plugin to register all others
